+++ deploy/common/k8sbased.sh
@@ -69,10 +69,6 @@ function launch_central {
         add_args "--monitoring-lb-type=$MONITORING_LOAD_BALANCER"
     fi
 
-    if [[ "$SCANNER_V2_SUPPORT" == "true" ]]; then
-        add_args "--enable-scanner-v2"
-    fi
-
     if [ -n "${OUTPUT_FORMAT}" ]; then
         add_args "--output-format=${OUTPUT_FORMAT}"
     fi
@@ -164,13 +160,7 @@ function launch_central {
         kubectl -n stackrox patch deploy/central --patch '{"spec":{"template":{"spec":{"containers":[{"name":"central","resources":{"limits":{"cpu":"1","memory":"4Gi"},"requests":{"cpu":"1","memory":"1Gi"}}}]}}}}'
     fi
 
-    if [[ "$SCANNER_V2_SUPPORT" == "true" ]]; then
-        echo "Deploying Scanner V2"
-        launch_service $unzip_dir scannerv2
-        echo
-        ${ORCH_CMD} -n stackrox patch deployment scanner-v2 --patch "$(cat $k8s_dir/scanner-v2-resources-patch.yaml)"
-        ${ORCH_CMD} -n stackrox patch deployment scanner-v2-db --patch "$(cat $k8s_dir/scanner-v2-db-resources-patch.yaml)"
-    elif [[ "$SCANNER_SUPPORT" == "true" ]]; then
+    if [[ "$SCANNER_SUPPORT" == "true" ]]; then
         echo "Deploying Scanning..."
         $unzip_dir/scanner/scripts/setup.sh
         launch_service $unzip_dir scanner